test(engine/world): エラー伝播テストを lib_test.go に統合する
InitGeneric の異常系検証を別ファイルに分けず、正常系と同じ lib_test.go に置く。
検証対象が同一なので1ファイルにまとめ、フェイク初期化子も同居させて追いやすくする。
